Divide by a value of the type the overload takes
The Linux build could not compile the calendars: an LL literal is a long long, and int64_t is
a long there rather than a long long, so a call with an LL literal on both sides matched none
of the overloads of FloorDiv::Divide exactly and was ambiguous between all three of them.

The arguments are now int64_t, which every platform agrees on. A call with one argument already
int64_t did resolve, because that argument picked the overload, but those are written out too
rather than left resting on it.

None of this can be seen on a machine where int64_t is a long long, which is why it was not
caught here. The failing calls compile under a header that mirrors the Linux types, and the
answers are unchanged: the zone sweep, the calendar tests and the property test over eighteen
calendars all give what they gave before.

+++ extension/icu/datetime/coptic.cpp
@@ -51,11 +51,11 @@ int64_t CopticEthiopicCalendar::HandleComputeMonthStart(int32_t eyear, int32_t m
 		month = month % 13 + 12;
 	}
 
-	return int64_t(GetEpochOffset())     // the day the first year starts on
-	       + 365LL * year                // the whole years
-	       + FloorDiv::Divide(year, 4LL) // their leap days
-	       + 30LL * month                // the whole months, which are all thirty days long
-	       - 1;                          // the day before the first of the month
+	return int64_t(GetEpochOffset())            // the day the first year starts on
+	       + 365 * year                         // the whole years
+	       + FloorDiv::Divide(year, int64_t(4)) // their leap days
+	       + 30 * month                         // the whole months, which are all thirty days long
+	       - 1;                                 // the day before the first of the month
 }
 
 void CopticEthiopicCalendar::HandleComputeFields(int32_t julian_day) {
